Adds query_func to BowlerTestCase.
In the beginning, we had `selector` and `modifier` arguments for this
function, but that didn't allow testing the built in selectors (like
`select_function`.  I added `selector_func` but this was a bit awkward
to use, and was never supported in `run_bowler_modifiers` (plural).

This is much less awkward, and allows you to do the full chain yourself:

    query_func = lambda x: Query(x).select_function("x").rename("y")

...and also captures exceptions better than the previous way.